requirements.txt
mss
numpy
PyQt6
pytesseract
Pillow
openai
transformers
python-dotenv